MANNING: bulk import AWS IAM users
MILESTONE 5

Instead of returning an error when a user already exists, we catch
the error code, set the username (later used in the Read function)
and pass on to the Read function.

After compiling the provider and copying this to the correct
Terraform plugins directory you are able to run a `terraform
apply` that bulk imports all users instead of erroring out.